+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   * Classes better organized.
   * Support [BSON](http://bsonspec.org) as serializer.
   * Optimize script ./bin/gen_dict_json.php allowing to create serialized dictionary files in selected formats.
+  * Created BSON/JSON/Msgpack files for the IDF file (file ./dict/idf.txt).
 
 0.25.2 / 2017-06-22
 ===================